refactor: extract github issue helper modules for #688
