+++ crates/tau-coding-agent/src/main.rs
@@ -18,6 +18,7 @@ mod events;
 mod extension_manifest;
 mod gemini_cli_client;
 mod github_issues;
+mod github_issues_helpers;
 mod macro_profile_commands;
 mod mcp_server;
 mod model_catalog;
